Addressing comments, moved feature utils into its own file
